requirements.txt
streamlit
google-generativeai
requests
Pillow
spotipy
matplotlib